tucana: Update powerhint from sunfish
powerhint.json: correct the path and value

Because the arch of sm7150 is different from 8150, correct the path and
adjust the value to align with sm7150. This is the initial version of
settings and still need to tuning for power/performance balance.

Bug: 142552311
Test: build and check the values
Change-Id: If061ee698c5f0164af876f171d504daa7a0c5847

powerhint.json: adjust GPU frequency by power level

We have 7150AA/7150AB devices at the same time and they have different
GPU freq table, so we use power level to adjust them

[Hlcpereira]: Sync powerhint.json with davinci

Test: powerhint test
Bug: 149962153
Change-Id: I3aa778eea762f1cd11d162a094db5c564269315e